Added some code to test the performance (max frequency/ticks per sec).
At the time, I got a very simple program (a slowed pixel-blinking loop)
to run with a CPU frequency of a bit over 300kHz. Of course, other
programs might have run slower (or faster), so it's just an estimate,
and it's not like the "screen" display actually updated that fast.