Replaced render-tick with UIManager, and made MemoryRenderer use it.
This means MemoryRenderer no longer listens to the RAM.write event, and
instead updates its UI when asked to refresh. It also triggers such a
UI refresh when memory is edited by the user.

This is a first step towards not using core-change events for updating
the UI, since at higher frequencies those happen far more often than
the UI should be updated, and at lower frequencies performance is not
quite as important (as it comparatively has plenty of time then).